include form validation for robot parameters

+++ robot_ui/robot/views.py
@@ -1,7 +1,11 @@
 from django.shortcuts import render, redirect
 from django.http import JsonResponse
 from django.views.decorators.csrf import csrf_exempt
-from django_htmx.http import trigger_client_event
+from django_htmx.http import (
+    trigger_client_event,
+    HttpResponseClientRedirect,
+    HttpResponseClientRefresh,
+)
 from .forms import PositionForm, RobotForm
 from .models import Position, Robot
 from .utils.jacobi import forward_kinematics
@@ -11,25 +15,28 @@ import json
 def main_view(request):
     objs = Position.objects.all()
     params = Robot.objects.all().first()
-    context = {"form": PositionForm(), "form_params": RobotForm(instance=params), "objs": objs, "params":params}
+    context = {
+        "form": PositionForm(),
+        "form_params": RobotForm(instance=params),
+        "objs": objs,
+        "params": params,
+    }
     return render(request, "index.html", context=context)
 
 
 def robot_parameters(request):
-    if request.POST:
-        form = RobotForm(data=request.POST)
-        if form.is_valid():
-            Robot.objects.all().delete()
-            obj = form.save(commit=False)
-            obj.save()
+    if request.htmx:
+        if request.POST:
+            form = RobotForm(data=request.POST)
+            if form.is_valid():
+                Robot.objects.all().delete()
+                obj = form.save(commit=False)
+                obj.save()
+                return HttpResponseClientRefresh()
 
-            return redirect("robot:main-view")
+            context = {"form_params": form}
 
-        objs = Position.objects.all()
-        context = {"form": PositionForm(), "form_params": form, "objs": objs}
-
-        return render(request, "index.html", context=context)
-
+            return render(request, "partials/form_params.html", context=context)
 
 
 def create_position(request):
